--- src/design-system/context/private/DesignSystemContext.ts
@@ -1,15 +1,0 @@
-import React from 'react';
-
-export type DesignSystemContextValue = {
-  experimentalFlags: {
-    androidShadowsV2?: boolean;
-  };
-};
-
-export const DesignSystemContext = React.createContext<DesignSystemContextValue>(
-  {
-    experimentalFlags: {
-      androidShadowsV2: false,
-    },
-  }
-);
