--- src/design-system/context/private/useExperimentalFlags.ts
@@ -1,7 +1,0 @@
-import { useContext } from 'react';
-
-import { DesignSystemContext } from './DesignSystemContext';
-
-export function useExperimentalFlags() {
-  return useContext(DesignSystemContext).experimentalFlags;
-}
